+++ packages/headless/src/api/knowledge/tests/stream-answer-api-state-mock.ts
@@ -31,6 +31,9 @@ export const streamAnswerAPIStateMock: StateNeededByAnswerAPI = {
       answerConfigurationId: '',
     },
   },
+  insightConfiguration: {
+    insightId: 'insight-id',
+  },
   version: '2.77.0',
   debug: false,
   pipeline: '',