+++ packages/samples/headless-ssr-commerce/app/_components/did-you-mean.tsx
@@ -1,1 +1,0 @@
-// TODO // TODO KIT-3463: implement did you mean in sample